+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Additional labels for pre-release and build metadata are available as extensions
 ## [2.X.0]
 1. Bug fix calling `is_file` function (Issue [#447](https://github.com/brightwind-dev/brightwind/issues/447)).
 2. Write new note here (Issue [#4xx](https://github.com/brightwind-dev/brightwind/issues/4xx)).
+3. Bug fix for `get_table` function (Issue [#421](https://github.com/brightwind-dev/brightwind/issues/421)).
 
 
 ## [2.2.1]